Add a shared demo package and move hd, amount, and ui into it.
