+++ demos/shared/README.md
@@ -1,0 +1,3 @@
+# mera demo shared code
+
+Shared code for the mera demos.
