+++ demos/shared/src/amount.ts
